Settings: move environment table to its own tab
