+++ rare/components/tabs/downloads/thread.py
@@ -1,8 +1,6 @@
-import multiprocessing
 import os
 import platform
 import queue
-import threading
 import time
 from dataclasses import dataclass
 from enum import IntEnum
